+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 ## Component Enhancements
 
 * `Message`: Two new props have been added, `border` and `roundedCorners`.
+* `Dropdown`: One new prop have been added, `cacheVisibleValue`.
 
 ## Helpers